updated bond logic, display, and added socket events.

+++ apps/backend/src/services/bond.service.ts
@@ -1,8 +1,9 @@
 // apps/backend/src/services/bond.service.ts
 
-import Bond, { IBond } from '../models/bond.model'
+import Bond from '../models/bond.model'
+import type { IBond } from '@iam/types'
 import { Types } from 'mongoose'
-import { HttpError } from '../utils/HttpError' // Import the HttpError
+import { HttpError } from '../utils/HttpError'
 
 interface BondData {
 	sender: Types.ObjectId | string
@@ -21,11 +22,16 @@ interface StatusUpdate {
  * @returns The created bond document
  */
 export const createBond = async (data: BondData): Promise<IBond> => {
+	if (!data.sender || !data.responder) {
+		throw new HttpError('Sender and responder are required', 400)
+	}
+
 	const bond = new Bond({
 		sender: data.sender,
 		responder: data.responder,
 		actionerId: data.sender,
 	})
+
 	return await bond.save()
 }
 
@@ -41,6 +47,10 @@ export const updateBondStatus = async (
 	statusUpdate: StatusUpdate,
 	actionerId: Types.ObjectId | string
 ): Promise<IBond> => {
+	if (!Types.ObjectId.isValid(id)) {
+		throw new HttpError('Invalid bond ID', 400)
+	}
+
 	const bond = await Bond.findById(id)
 	if (!bond) throw new HttpError('Bond not found', 404)
 
@@ -56,18 +66,28 @@ export const updateBondStatus = async (
 export const getUserBonds = async (
 	userId: Types.ObjectId | string
 ): Promise<IBond[]> => {
+	if (!Types.ObjectId.isValid(userId.toString())) {
+		throw new HttpError('Invalid user ID', 400)
+	}
+
 	return await Bond.find({
 		$or: [{ sender: userId }, { responder: userId }],
 	})
-	// .populate('sender responder')
 }
 
 /**
  * Deletes a bond by ID.
  * @param id - The bond document ID
  */
-export const deleteBond = async (id: string): Promise<void> => {
+export const deleteBond = async (id: string): Promise<IBond> => {
+	if (!Types.ObjectId.isValid(id)) {
+		throw new HttpError('Invalid bond ID', 400)
+	}
 	const bond = await Bond.findById(id)
 	if (!bond) throw new HttpError('Bond not found', 404)
+	
+	const deletedBond = bond
 	await bond.deleteOne()
+
+	return deletedBond
 }
